--- x/asset/nft/client/cli/global_freeze_test.go
@@ -1,61 +1,0 @@
-package cli_test
-
-import (
-	"testing"
-
-	clitestutil "github.com/cosmos/cosmos-sdk/testutil/cli"
-	"github.com/google/uuid"
-	"github.com/stretchr/testify/require"
-
-	"github.com/CoreumFoundation/coreum/app"
-	"github.com/CoreumFoundation/coreum/pkg/config"
-	"github.com/CoreumFoundation/coreum/pkg/config/constant"
-	"github.com/CoreumFoundation/coreum/testutil/network"
-	"github.com/CoreumFoundation/coreum/x/asset/client/cli"
-	"github.com/CoreumFoundation/coreum/x/asset/types"
-)
-
-func TestGloballyFreezeUnfreezeFungibleToken(t *testing.T) {
-	requireT := require.New(t)
-	networkCfg, err := config.NetworkByChainID(constant.ChainIDDev)
-	requireT.NoError(err)
-	app.ChosenNetwork = networkCfg
-	testNetwork := network.New(t)
-
-	// the denom must start from the letter
-	symbol := "l" + uuid.NewString()[:4]
-	subunit := "sub" + symbol
-	precision := "8"
-	ctx := testNetwork.Validators[0].ClientCtx
-	issuer := testNetwork.Validators[0].Address
-	denom := types.BuildFungibleTokenDenom(subunit, issuer)
-
-	// Issue token
-	args := []string{symbol, subunit, precision, "777", `"My Token"`,
-		"--features", types.FungibleTokenFeature_freeze.String(), //nolint:nosnakecase
-	}
-	args = append(args, txValidator1Args(testNetwork)...)
-	_, err = clitestutil.ExecTestCLICmd(ctx, cli.CmdTxIssueFungibleToken(), args)
-	requireT.NoError(err)
-
-	// Globally freeze the token
-	args = append([]string{denom, "--output", "json"}, txValidator1Args(testNetwork)...)
-	_, err = clitestutil.ExecTestCLICmd(ctx, cli.CmdTxGloballyFreezeFungibleToken(), args)
-	requireT.NoError(err)
-
-	var resp types.QueryFungibleTokenResponse
-	buf, err := clitestutil.ExecTestCLICmd(ctx, cli.CmdQueryFungibleToken(), []string{denom, "--output", "json"})
-	requireT.NoError(err)
-	requireT.NoError(ctx.Codec.UnmarshalJSON(buf.Bytes(), &resp))
-	requireT.True(resp.FungibleToken.GloballyFrozen)
-
-	// Globally unfreeze the token
-	args = append([]string{denom, "--output", "json"}, txValidator1Args(testNetwork)...)
-	_, err = clitestutil.ExecTestCLICmd(ctx, cli.CmdTxGloballyUnfreezeFungibleToken(), args)
-	requireT.NoError(err)
-
-	buf, err = clitestutil.ExecTestCLICmd(ctx, cli.CmdQueryFungibleToken(), []string{denom, "--output", "json"})
-	requireT.NoError(err)
-	requireT.NoError(ctx.Codec.UnmarshalJSON(buf.Bytes(), &resp))
-	requireT.False(resp.FungibleToken.GloballyFrozen)
-}
